LPD-26955 Move logic from LayoutPage.ts to reuse existent PagesAdminPage.ts
