+++ modules/test/playwright/tests/layout-admin-web/fixtures/pagesPagesTest.ts
@@ -5,22 +5,17 @@
 
 import {test} from '@playwright/test';
 
-import {LayoutPage} from '../pages/LayoutPage';
 import {PageConfigurationPage} from '../pages/PageConfigurationPage';
 import {StaticPagesPage} from '../pages/StaticPagesPage';
 import {UtilityPageConfigurationPage} from '../pages/UtilityPageConfigurationPage';
 import {UtilityPagesPage} from '../pages/UtilityPagesPage';
 
 const pagesPagesTest = test.extend<{
-	layoutPage: LayoutPage;
 	pageConfigurationPage: PageConfigurationPage;
 	staticPagesPage: StaticPagesPage;
 	utilityPageConfigurationPage: UtilityPageConfigurationPage;
 	utilityPagesPage: UtilityPagesPage;
 }>({
-	layoutPage: async ({page}, use) => {
-		await use(new LayoutPage(page));
-	},
 	pageConfigurationPage: async ({page}, use) => {
 		await use(new PageConfigurationPage(page));
 	},
