--- modules/test/playwright/tests/layout-admin-web/pages/LayoutPage.ts
@@ -1,36 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {Page} from '@playwright/test';
-
-import {clickAndExpectToBeVisible} from '../../../utils/clickAndExpectToBeVisible';
-import {PORTLET_URLS} from '../../../utils/portletUrls';
-
-export class LayoutPage {
-	readonly page: Page;
-
-	constructor(page: Page) {
-		this.page = page;
-	}
-
-	async goto(siteUrl?: Site['friendlyUrlPath']) {
-		await this.page.goto(
-			`/group${siteUrl || '/guest'}${PORTLET_URLS.pages}`
-		);
-	}
-
-	async clickOnAction(action: string, title: string) {
-		await clickAndExpectToBeVisible({
-			autoClick: true,
-			target: this.page.getByRole('menuitem', {
-				exact: true,
-				name: action,
-			}),
-			trigger: this.page
-				.locator('li', {has: this.page.getByText(title)})
-				.getByRole('button', {name: 'Open Page Options Menu'}),
-		});
-	}
-}
